+++ frontend/src/components/ui/StationDetailsWindow/reports/charger/ChargerReportConfirmation.tsx
