+++ frontend/src/components/ui/StationDetailsWindow/reports/station/StationReportConfirmation.tsx
@@ -11,7 +11,7 @@ import FlexBox from '@common/FlexBox';
 import Text from '@common/Text';
 import TextField from '@common/TextField';
 
-import { findDifferentKeys } from '@ui/StationDetailsWindow/reports/domain';
+import { findDifferentKeys } from '@ui/StationDetailsWindow/reports/station/domain';
 import StationInformation from '@ui/StationDetailsWindow/station/StationInformation';
 
 import {
