+++ frontend/src/components/ui/StationDetailsWindow/reports/station/domain.ts
